+++ node/libs/roles/src/validator/messages/leader_commit.rs
@@ -45,9 +45,9 @@ pub enum CommitQCVerifyError {
     #[error("Signers have not reached wanted weight: got {got}, want {want}")]
     WeightNotReached {
         /// Got weight.
-        got: usize,
+        got: u64,
         /// Want weight.
-        want: usize,
+        want: u64,
     },
     /// Bad signature.
     #[error("bad signature: {0:#}")]
@@ -100,7 +100,7 @@ impl CommitQC {
             return Err(Error::BadSignersSet);
         }
 
-        // Verify the signer's weights is enough.
+        // Verify the signers' weight is enough.
         let weight = genesis.validators.weight_from_signers(self.signers.clone());
         let threshold = genesis.validators.threshold();
         if weight < threshold {
